Rename webjars-locator to web-dependency-locator
