+++ devtools/bom-descriptor-json/pom.xml
@@ -2867,7 +2867,7 @@
                 </dependency>
                 <dependency>
                     <groupId>io.quarkus</groupId>
-                    <artifactId>quarkus-webjars-locator</artifactId>
+                    <artifactId>quarkus-web-dependency-locator</artifactId>
                     <version>${project.version}</version>
                     <type>pom</type>
                     <scope>test</scope>
