+++ docs/src/main/asciidoc/http-reference.adoc
@@ -10,8 +10,9 @@ include::_attributes.adoc[]
 :numbered:
 :sectnums:
 :sectnumlevels: 4
-:topics: http,web,webjars,vertx,servlet,undertow
+:topics: http,web,webjars,mvnpm,vertx,servlet,undertow
 :extensions: io.quarkus:quarkus-vertx-http
+:web-locator-ga: quarkus-web-dependency-locator
 
 This document clarifies different HTTP functionalities available in Quarkus.
 
@@ -32,39 +33,10 @@ was chosen as it is the standard location for resources in `jar` files as define
 Quarkus can be used without Servlet, following this convention allows existing code that places its resources in this
 location to function correctly.
 
-[[from-mvnpm]]
-=== From mvnpm
+=== From web dependencies like webjars or mvnpm
 
-If you are using https://mvnpm.org/[mvnpm], as for the following JQuery dependency:
-
-[source,xml,role="primary asciidoc-tabs-target-sync-cli asciidoc-tabs-target-sync-maven"]
-.pom.xml
-----
-<dependency>
-    <groupId>org.mvnpm</groupId>
-    <artifactId>bootstrap</artifactId>
-    <version>5.3.3</version>
-    <scope>runtime</scope>
-</dependency>
-----
-
-[source,gradle,role="secondary asciidoc-tabs-target-sync-gradle"]
-.build.gradle
-----
-runtimeOnly("org.mvnpm:bootstrap:5.3.3")
-----
-
-You can import it in your HTML like this:
-[source,html]
-----
-<script src="_static/bootstrap/5.3.3/dist/css/bootstrap.min.css"></script>
-----
-
-
-[[from-webjars]]
-=== From WebJars
-
-If you are using webjars, like the following JQuery one:
+==== WebJars
+If you are using https://www.webjars.org[WebJars], like the following JQuery one:
 
 [source,xml,role="primary asciidoc-tabs-target-sync-cli asciidoc-tabs-target-sync-maven"]
 .pom.xml
@@ -83,7 +55,7 @@ implementation("org.webjars:jquery:3.1.1")
 ----
 
 and rather write `/webjars/jquery/jquery.min.js` instead of `/webjars/jquery/3.1.1/jquery.min.js`
-in your HTML files, you can add the `quarkus-webjars-locator` extension to your project.
+in your HTML files, you can add the `{web-locator-ga}` extension to your project.
 To use it, add the following to your project's dependencies:
 
 [source,xml,role="primary asciidoc-tabs-target-sync-cli asciidoc-tabs-target-sync-maven"]
@@ -91,15 +63,65 @@ To use it, add the following to your project's dependencies:
 ----
 <dependency>
     <groupId>io.quarkus</groupId>
-    <artifactId>quarkus-webjars-locator</artifactId>
+    <artifactId>{web-locator-ga}</artifactId>
 </dependency>
 ----
 
 [source,gradle,role="secondary asciidoc-tabs-target-sync-gradle"]
 .build.gradle
 ----
-implementation("io.quarkus:quarkus-webjars-locator")
+implementation("io.quarkus:{web-locator-ga}")
 ----
+
+==== Mvnpm
+
+If you are using https://mvnpm.org[mvnpm], like the following Lit one:
+
+[source,xml,role="primary asciidoc-tabs-target-sync-cli asciidoc-tabs-target-sync-maven"]
+.pom.xml
+----
+<dependency>
+    <groupId>org.mvnpm</groupId>
+    <artifactId>lit</artifactId>
+    <version>3.1.2</version>
+</dependency>
+----
+
+[source,gradle,role="secondary asciidoc-tabs-target-sync-gradle"]
+.build.gradle
+----
+implementation("org.mvnpm:lit:3.1.2")
+----
+
+you can use the `{web-locator-ga}` as described above to reference the resource without the version, however with mvnpm you can
+also use https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script/type/importmap[importmaps].
+
+The importmap is generated by the `{web-locator-ga}` extension, and available at `/_importmap/generated_importmap.js`.
+This mean adding the following to your `index.html` will allow you to import web libraries by name:
+
+[source,html]
+----
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>My app</title>
+
+    <script src="/_importmap/generated_importmap.js"></script> <1>
+
+    <script type="module">
+        import '@lit'; <2>
+        import 'app/demo-app.js'; <3>
+    </script>
+  </head>
+    
+</html>
+----
+<1> Use the generated importmap
+<2> Import web libraries
+<3> Import your own files, this can be done by adding `quarkus.web-dependency-locator.import-mappings.app/ = /app/` to the config. Any key-value pair can be added.
+
 
 === From a local directory
