+++ extensions/web-dependency-locator/deployment/src/test/java/io/quarkus/webdependency/locator/test/WebDependencyLocatorTestSupport.java
@@ -1,0 +1,8 @@
+package io.quarkus.webdependency.locator.test;
+
+class WebDependencyLocatorTestSupport {
+
+    static final String JQUERY_UI_VERSION = System.getProperty("webjar.jquery-ui.version");
+    static final String MOMENTJS_VERSION = System.getProperty("webjar.momentjs.version");
+    static final String BOOTSTRAP_VERSION = System.getProperty("mvnpm.bootstrap.version");
+}
