--- extensions/webjars-locator/deployment/src/main/java/io/quarkus/webjar/locator/deployment/ImportMapBuildItem.java
@@ -1,15 +1,0 @@
-package io.quarkus.webjar.locator.deployment;
-
-import io.quarkus.builder.item.SimpleBuildItem;
-
-public final class ImportMapBuildItem extends SimpleBuildItem {
-    private final String importmap;
-
-    public ImportMapBuildItem(String importmap) {
-        this.importmap = importmap;
-    }
-
-    public String getImportMap() {
-        return this.importmap;
-    }
-}
