--- extensions/webjars-locator/deployment/src/test/java/io/quarkus/webjar/locator/test/WebJarLocatorTestSupport.java
@@ -1,8 +1,0 @@
-package io.quarkus.webjar.locator.test;
-
-class WebJarLocatorTestSupport {
-
-    static final String JQUERY_UI_VERSION = System.getProperty("webjar.jquery-ui.version");
-    static final String MOMENTJS_VERSION = System.getProperty("webjar.momentjs.version");
-    static final String BOOTSTRAP_VERSION = System.getProperty("mvnpm.bootstrap.version");
-}
