--- extensions/webjars-locator/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,12 +1,0 @@
----
-artifact: ${project.groupId}:${project.artifactId}:${project.version}
-name: "WebJar Locator"
-metadata:
-  short-name: "webjars-locator"
-  keywords:
-  - "web"
-  - "webjar"
-  guide: "https://quarkus.io/guides/http-reference#webjar-locator-support"
-  categories:
-  - "web"
-  status: "stable"
